requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gspread==5.12.0
oauth2client==4.1.3
openai==1.12.0
python-docx==1.1.0
PyPDF2==3.0.1
